Iniciando aplicação Flask

+++ sql/Procedures.sql
@@ -1,3 +1,5 @@
+-- Procedure Verifica se a Tarefa foi Encerrada
+
 -- Procedure Inserir
 CREATE PROCEDURE ListarTarefas
 AS
@@ -5,6 +7,7 @@ BEGIN
     SELECT *
     FROM Tarefas
 END
+GO
 
 -- Procedure Buscar Status atual
 CREATE PROCEDURE BuscarStatus
@@ -15,6 +18,7 @@ BEGIN
     FROM Tarefas
     WHERE TarefaId = @TarefaId
 END
+GO
 
 -- Procedure Inserir
 CREATE PROCEDURE AdicionarNovaTarefa
@@ -26,28 +30,61 @@ BEGIN
     VALUES
         (@Descricao, 'Não Iniciada', GETDATE())
 END
+GO
 
 -- Procedure Atualizar Status
 CREATE PROCEDURE AtualizarStatusTarefa
     @Status NVARCHAR(50),
     @TarefaId INT
 AS
-BEGIN
-    UPDATE Tarefas set Status = @Status WHERE TarefaId = @TarefaId
-END
+IF EXISTS(SELECT 1 FROM Tarefas WHERE TarefaId = @TarefaId AND Status != 'Concluída' AND Status != 'Cancelada')
+    BEGIN
+        UPDATE Tarefas set Status = @Status WHERE TarefaId = @TarefaId
+    END
+ELSE
+    BEGIN
+        PRINT 'Tarefa já foi encerrada. Atualização não permitida.'
+    END
+GO
 
 -- Procedure Tarefa Concluida
 CREATE PROCEDURE ConcluirTarefa
     @TarefaId INT
 AS
-BEGIN
-    UPDATE Tarefas set Status = 'Concluído', DataConclusao = GETDATE() WHERE TarefaId = @TarefaId
-END
+IF EXISTS(SELECT 1 FROM Tarefas WHERE TarefaId = @TarefaId AND Status != 'Concluída' AND Status != 'Cancelada')
+    BEGIN
+        UPDATE Tarefas set Status = 'Concluída', DataConclusao = GETDATE() WHERE TarefaId = @TarefaId
+    END
+ELSE
+    BEGIN
+        PRINT 'Tarefa já está encerrada. Atualização não permitida.'
+    END
+GO
 
 -- Procedure Excluir Tarefa
 CREATE PROCEDURE DeletarTarefa
     @TarefaId INT
 AS
-BEGIN
-    DELETE FROM Tarefas WHERE TarefaId = @TarefaId
-END
+IF EXISTS(SELECT 1 FROM Tarefas WHERE TarefaId = @TarefaId AND Status != 'Concluída' AND Status != 'Cancelada')
+    BEGIN
+        DELETE FROM Tarefas WHERE TarefaId = @TarefaId
+    END
+ELSE
+    BEGIN
+        PRINT 'Tarefa já está encerrada. Atualização não permitida.'
+    END
+GO
+
+-- Procedure Tarefa Cancelada
+CREATE PROCEDURE CancelarTarefa
+    @TarefaId INT
+AS
+IF EXISTS(SELECT 1 FROM Tarefas WHERE TarefaId = @TarefaId AND Status != 'Concluída' AND Status != 'Cancelada')
+    BEGIN
+        UPDATE Tarefas set Status = 'Cancelada', DataConclusao = GETDATE() WHERE TarefaId = @TarefaId
+    END
+ELSE
+    BEGIN
+        PRINT 'Tarefa já está encerrada. Atualização não permitida.'
+    END
+GO
